Move enemyMoves to utils
